% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood_helpers.R
\name{check_compat_log_lik}
\alias{check_compat_log_lik}
\title{# Check compatibility of all parameters of log-likelihood function}
\usage{
check_compat_log_lik(Beta, b_vec, lambda_vec)
}
\arguments{
\item{Beta}{Numeric vector or mxp matrix; represents m different weight vectors for length p cosine basis; must have Euclidean norm less than pi.}

\item{b_vec}{Numeric vector; input points to interpolation; must be in order, first element equal to 0 and second element equal to 1.}

\item{lambda_vec}{Numeric vector; output points to interpolation; must have lambda_vec[i] > lambda_vec[i] < lambda_vec[i + 1].}
}
\value{
modified beta matrix in case some needed to be removed
}
\description{
# Check compatibility of all parameters of log-likelihood function
}
