% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood_helpers.R
\name{check_fluctuating_lambda}
\alias{check_fluctuating_lambda}
\title{Check if in lambda vector lambda_{i - 1} > lambda{i} < lambda_{i + 1} when i is odd (lambda_vec length 5 or higher)}
\usage{
check_fluctuating_lambda(lambda_vec)
}
\arguments{
\item{lambda_vec}{vector of positive numbers}
}
\description{
Check if in lambda vector lambda_{i - 1} > lambda{i} < lambda_{i + 1} when i is odd (lambda_vec length 5 or higher)
}
