% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_mcmc_helpers.R
\name{check_integer_2}
\alias{check_integer_2}
\title{Check if positive integer greater than or equal to 2}
\usage{
check_integer_2(x, name)
}
\arguments{
\item{x}{number to check}

\item{name}{variable name of number to check}
}
\description{
Check if positive integer greater than or equal to 2
}
