% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood_helpers.R
\name{check_matrix}
\alias{check_matrix}
\title{Check if input is matrix. If not, turn into 1 column matrix}
\usage{
check_matrix(X)
}
\arguments{
\item{X}{array like}
}
\value{
matrix form of X (nx1)
}
\description{
Check if input is matrix. If not, turn into 1 column matrix
}
