% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_mcmc_helpers.R
\name{check_pos_integer}
\alias{check_pos_integer}
\title{Check if positive integer}
\usage{
check_pos_integer(x, name)
}
\arguments{
\item{x}{number to check}

\item{name}{variable name of number to check}
}
\description{
Check if positive integer
}
