% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood_helpers.R
\name{cos_2_integral}
\alias{cos_2_integral}
\title{Calculate int_{0}^{x}cos^2(y * pi * j)dy}
\usage{
cos_2_integral(x, j)
}
\arguments{
\item{x}{number, upper bound}

\item{j}{positive integer}
}
\value{
integral result
}
\description{
Calculate int_{0}^{x}cos^2(y * pi * j)dy
}
