% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood_helpers.R
\name{cos_cos_integral}
\alias{cos_cos_integral}
\title{Calculate int_{0}^{x} cos(y * pi * j1)cos(y * pi * j2)dy where j1 != j2}
\usage{
cos_cos_integral(x, j1, j2)
}
\arguments{
\item{x}{number, upper bound}

\item{j1}{positive integer}

\item{j2}{positive integer, different from j1}
}
\value{
integral result
}
\description{
Calculate int_{0}^{x} cos(y * pi * j1)cos(y * pi * j2)dy where j1 != j2
}
