% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood_helpers.R
\name{cosine_basis}
\alias{cosine_basis}
\title{Given weights (Beta), calculates cosine basis evaluated at X}
\usage{
cosine_basis(X, Beta)
}
\arguments{
\item{X}{nx1 Matrix of observations}

\item{Beta}{mxj Matrix of different beta vectors (weight parameters)}
}
\value{
nxm Matrix of representing the n observations evaluated at m different cosine bases
}
\description{
Given weights (Beta), calculates cosine basis evaluated at X
}
