% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffeo_bayes_estimation.R
\name{diffeo_bayes_estimate}
\alias{diffeo_bayes_estimate}
\title{Estimate Distribution and Obtain Posterior of Mode with Diffeomorphism Bayesian Analysis}
\usage{
diffeo_bayes_estimate(
  X,
  num_betas,
  num_samples = 5000,
  burn_in = 1000,
  prior_sd = 0.75,
  prop_sd = 0.5,
  plot_results = TRUE
)
}
\arguments{
\item{X}{Numeric vector; represents data vector}

\item{num_betas}{positive integer > 1; length of weight vector, also number of basis elements for cosine basis of diffeomorphism.}

\item{num_samples}{positive integer; number of samples to return from MCMC.}

\item{burn_in}{positive integer; burn_in number for MCMC.}

\item{prior_sd}{positive real; prior is no covariance multivariate normal, represents standard deviation for each component}

\item{prop_sd}{positive real; proposal distribution standard deviation for adaptive MCMC}
}
\value{
List containing following components: \item{sampled_betas}{MCMC posterior sampled weight vectors} \item{sampled_modes}{MCMC posterior sampled modes} \item{bayes_map_beta}{MAP estimate of beta vector} \item{bayes_map_mode}{MAP estimate of mode} \item{p_X}{the input used for pdf estimation} \item{bayes_map_pdf}{MAP estimate of the pdf}
}
\description{
Obtain MAP estimates of pdf and posterior distribution of modes by calculating posterior with adapative MCMC
}
\examples{
# Sample from gamma(2, 1). Mode should be around 1
# X <- rgamma(50, 2, 1)
# bayes_estimate <- diffeo_bayes_estimate(X, num_betas = 3, num_samples = 5000, prior_sd = 0.75)
}
