% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{diffeo_bayes_plot}
\alias{diffeo_bayes_plot}
\title{Plot for bayesian analysis of modes. Displays mode estimation plot with MAP estimates on top and histogram of MCMC sampled modes on bottom}
\usage{
diffeo_bayes_plot(X, p_X, pdf, mode_estimate, sampled_modes)
}
\arguments{
\item{X}{numeric vector; original data}

\item{p_X}{numeric vector; data used for pdf estimation}

\item{pdf}{numeric vector; pdf output with p_X as input}

\item{mode_estimate}{real number; estimated mode, to be displayed on top left or top right corner}

\item{sampled_modes}{numeric vector; contains sampled modes from Bayesian MCMC}
}
\description{
Plot for bayesian analysis of modes. Displays mode estimation plot with MAP estimates on top and histogram of MCMC sampled modes on bottom
}
