% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffeo_log_likelihood.R
\name{diffeo_log_likelihood}
\alias{diffeo_log_likelihood}
\title{Interpolation Log-Likelihood with Diffeomorphisms}
\usage{
diffeo_log_likelihood(
  X,
  Beta,
  b_vec,
  lambda_vec,
  interpolation = "cubic",
  transform_X = TRUE,
  check_compat = TRUE
)
}
\arguments{
\item{X}{Numeric vector or nx1 matrix; represents data vector.}

\item{Beta}{Numeric vector or mxp matrix; represents m different weight vectors for length p cosine basis; must have Euclidean norm less than pi.}

\item{b_vec}{Numeric vector; input points to interpolation; must be in order, first element equal to 0 and second element equal to 1.}

\item{lambda_vec}{Numeric vector; output points to interpolation; must have lambda_vec[i] > lambda_vec[i] < lambda_vec[i + 1].}

\item{interpolation}{'cubic' or 'linear'; the interpolation method to use to fit b_vec & lambda_vec.}

\item{transform_X}{boolean; whether to transform X into values in between 0 and 1 using min-max scaling; will do so automatically if min(X) < 0 or max(X) > 1.}

\item{check_compat}{boolean; whether to perform process to check compatibility of given parameters.}
}
\value{
List with following components: \item{diffeomorphism}{nxm matrix containing output after applying differomorphism to X for each beta weight vector} \item{log_like}{nxm matrix containing output after applying log-likelihood to X for each beta weight vector}
}
\description{
Calculates log-likelihood and diffeomorphism map for a set of points
}
\examples{
# 3 modes (5 extrema) with two different beta weight vectors (-0.4, 0.3, 0.7) & (0.6, -0.1, 0.2)
X <- runif(100, 0, 1)
Beta <- matrix(c(-0.4, 0.3, 0.7, 0.6, -0.1, 0.2), nrow = 2, byrow = TRUE)
lambda_vec <- c(0, 1, 0.2, 0.4, 0.3, 2, 0)
b_vec <- seq(0, 1, length.out = length(lambda_vec))
result_lst <- diffeo_log_likelihood(X, Beta, b_vec, lambda_vec)
}
