% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffeo_mle_estimation.R
\name{diffeo_mle_estimate}
\alias{diffeo_mle_estimate}
\title{Estimate Distribution and its Mode with Diffeomorphism MLE}
\usage{
diffeo_mle_estimate(
  X,
  num_betas,
  num_trials = 25,
  beta_starts = NULL,
  plot_results = TRUE
)
}
\arguments{
\item{X}{Numeric vector; represents data vector}

\item{num_betas}{positive integer > 1; length of weight vector, also number of basis elements for cosine basis of diffeomorphism.}

\item{num_trials}{positive integer; number of random starting points, and thus trials, for optimization procedure.}

\item{beta_starts}{NULL or n x num_betas matrix; represents user chosen starting points. If NULL num_trials random beta vectors chosen.}

\item{plot_results}{boolean; whether to display plot which shows histogram of original data, the pdf estimate and the mode estimate.}
}
\value{
List with following components: \item{final_beta}{MLE weight estimate} \item{mode_estimate}{MLE mode estimate} \item{p_X}{input used for MLE pdf estimation} \item{mle_pdf}{MLE pdf estimate with p_X}
}
\description{
Obtain points estimates of pdf and mode by calculating MLE of diffeomorphism log-likelihood
}
\examples{
# Sample from gamma(2, 1). Mode should be around 1
X <- rgamma(50, 2, 1)
mle_estimate <- diffeo_mle_estimate(X, num_betas = 3, num_trials = 25, plot_results = TRUE)
}
