% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood_helpers.R
\name{different_beta_product}
\alias{different_beta_product}
\title{Given vector, output 2 x choose(length(vec), 2) containing every combination of elements within vec}
\usage{
different_beta_product(vec)
}
\arguments{
\item{vec}{1-d vector}
}
\value{
vec containing every possible product of two numbers
}
\description{
Given vector, output 2 x choose(length(vec), 2) containing every combination of elements within vec
}
