% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdiff-package.R
\docType{package}
\name{exdiff-package}
\alias{exdiff}
\alias{exdiff-package}
\title{exdiff: Extrema Inference via Diffeomorphisms for Multimodial Probability Distributions}
\description{
This packages offers functionalities to estimate the extrema of probability distributions. By applying a certain class of increasing, continuous functions called diffeomorphisms to a simple class of spline interpolations, any probability distribution within a specified compact interval and a given number of extrema, or nodes, can be estimated. In addition, with our formulation, the extrema can easily be inferred after the estimation procedure is complete. Currently, we offer both a maximum likelohood functionality that returns point estimates of the mode for unimodal distributions and a Bayesian implementation via an adaptive MCMC procedure that outputs posterior samples of the mode for unimodal distributions. Finally, we offer a graphical implementation that displays both the estimated probability distribution and the location of the mode.
}
\keyword{internal}
