% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood_helpers.R
\name{gamma_function}
\alias{gamma_function}
\title{Gamma function which transforms domain (0, 1) to new domain (0, 1) via diffeomorphism}
\usage{
gamma_function(X, Beta)
}
\arguments{
\item{X}{nx1 Matrix of observations}

\item{Beta}{mxj Matrix of different beta vectors (weight parameters)}
}
\value{
nxm Matrix of representing the results of the gamma function defined in the paper (int_{0}^{X} exponential_map(t)^2 dt)
}
\description{
Gamma function which transforms domain (0, 1) to new domain (0, 1) via diffeomorphism
}
