% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_mcmc_helpers.R
\name{global_optimize}
\alias{global_optimize}
\title{Given dataset, find optimal beta vector that maximizes MLE or MAP (Only assumes one mode, so designed to find most extreme mode in probability distribution) Uses fmincon with multiple starting points and picks}
\usage{
global_optimize(
  input_X,
  num_betas,
  optimize_type,
  num_trials = 25,
  beta_starts = NULL,
  prior_sd = NULL
)
}
\arguments{
\item{input_X}{numeric vector; represents data for log likelihood. Points must be in between 0 and 1}

\item{num_betas}{positive integer; length of weight vector, also number of basis elements for cosine basis of diffeomorphism}

\item{optimize_type}{either 'mle' or 'map', whether to find mle or map estimate}

\item{num_trials}{positive integer; number of random starting points for fmincon}

\item{beta_starts}{- NULL or n x num_betas matrix; represents user chosen starting points. If NULL num_trials random beta vectors chosen}

\item{prior_sd}{positive real or NULL; Represents prior distribution standard deviation for MAP}
}
\value{
numeric vector representing fmincon's estimate of the argmin should be
}
\description{
Given dataset, find optimal beta vector that maximizes MLE or MAP (Only assumes one mode, so designed to find most extreme mode in probability distribution) Uses fmincon with multiple starting points and picks
}
