% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_mcmc_helpers.R
\name{hin1}
\alias{hin1}
\title{Function used for fmincon. Makes sure norm of betas < pi}
\usage{
hin1(x)
}
\arguments{
\item{x}{vector}
}
\value{
difference of Euclidean norm^2 and pi^2
}
\description{
Function used for fmincon. Makes sure norm of betas < pi
}
