% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood_helpers.R
\name{int_cosine_basis_2}
\alias{int_cosine_basis_2}
\title{Calculate integral from 0 to X of cosine basis squared with beta as weights}
\usage{
int_cosine_basis_2(X, Beta)
}
\arguments{
\item{X}{nx1 Matrix of observations}

\item{Beta}{mxj Matrix of different beta vectors (weight parameters)}
}
\value{
nxm Matrix of representing the integral of 0 to X of cosine basis squared with different weight parameters
}
\description{
Calculate integral from 0 to X of cosine basis squared with beta as weights
}
