% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_mcmc_helpers.R
\name{mh_adaptive_sampling_betas}
\alias{mh_adaptive_sampling_betas}
\title{For Bayesian estimation, runs adaptive MCMC to get posterior distribution of beta_vectors}
\usage{
mh_adaptive_sampling_betas(
  input_X,
  map_estimate,
  num_samples = 5000,
  burn_in = 1000,
  prior_sd = 0.75,
  prop_sd = 0.5
)
}
\arguments{
\item{input_X}{numeric vector; represents data for log likelihood. Points must be in between 0 and 1}

\item{map_estimate}{NULL numeric vector; MAP estimate which is starting point for MCMC}

\item{num_samples}{positive integer; number of samples to return from MCMC}

\item{burn_in}{positive integer; burn_in number for MCMC}

\item{prior_sd}{positive real; prior is no covariance multivariate normal, represents standard deviation for each component}

\item{prop_sd}{positive real; proposal distribution standard deviation for adaptive MCMC}
}
\value{
list containing sampled beta vectors (sampled_betas), sampled modes calculated from the sampled beta vectors (sampled_modes), MAP estimate of the mode (bayes_map_mode), and MAP estimate of the pdf using X_SAMPLE as input (bayes_map_pdf)
}
\description{
For Bayesian estimation, runs adaptive MCMC to get posterior distribution of beta_vectors
}
