% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood_helpers.R
\name{min_max_transform_X}
\alias{min_max_transform_X}
\title{Helper to transform X to interval in between 0 and 1 (Required for diffeomorphism)}
\usage{
min_max_transform_X(X)
}
\arguments{
\item{X}{vector or matrix containing data}
}
\value{
list containing transformed X (new_X), lower_bound used in transformation and upper bound used in transformation
}
\description{
Helper to transform X to interval in between 0 and 1 (Required for diffeomorphism)
}
