% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_mcmc_helpers.R
\name{mle_optimization}
\alias{mle_optimization}
\title{Given beta (weight) vector and dataset with values in between 0 and 1, calculate negative joint log likelihood (Only one mode).}
\usage{
mle_optimization(betas, input_X)
}
\arguments{
\item{betas}{numeric vector; represents weight vector for cosine basis}

\item{input_X}{numeric vector; represents data for log likelihood. Points must be in between 0 and 1}
}
\value{
negative log likelihood
}
\description{
Given beta (weight) vector and dataset with values in between 0 and 1, calculate negative joint log likelihood (Only one mode).
}
