% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{mode_estimation_plot}
\alias{mode_estimation_plot}
\title{Displays histogram of original data, pdf estimate and mode estimate}
\usage{
mode_estimation_plot(X, p_X, pdf, mode_estimate, plot_title)
}
\arguments{
\item{X}{numeric vector; original data}

\item{p_X}{numeric vector; data used for pdf estimation}

\item{pdf}{numeric vector; pdf output with p_X as input}

\item{mode_estimate}{real number; estimated mode, to be displayed on top left or top right corner}

\item{plot_title}{string; title of plot}
}
\description{
Displays histogram of original data, pdf estimate and mode estimate
}
