% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood_helpers.R
\name{TRAPZ_INTEGRAL_FUNC}
\alias{TRAPZ_INTEGRAL_FUNC}
\title{Calculating integral from 0 to 1 of function via trapezoid integral}
\usage{
TRAPZ_INTEGRAL_FUNC(func)
}
\arguments{
\item{func}{function to integrate}
}
\value{
integral result
}
\description{
Calculating integral from 0 to 1 of function via trapezoid integral
}
