% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood_helpers.R
\name{turn_1d_into_matrix}
\alias{turn_1d_into_matrix}
\title{Check if input is matrix. If not, turn into 1 row matrix}
\usage{
turn_1d_into_matrix(X)
}
\arguments{
\item{X}{array like}
}
\value{
matrix form of X (1xn)
}
\description{
Check if input is matrix. If not, turn into 1 row matrix
}
