% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood_helpers.R
\docType{data}
\name{X_FOR_INTEGRAL}
\alias{X_FOR_INTEGRAL}
\title{Used for calculating normalizing constant (Trapezoid integration)}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2000 rows and 1 columns.
}
\usage{
X_FOR_INTEGRAL
}
\description{
Used for calculating normalizing constant (Trapezoid integration)
}
\keyword{datasets}
