% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_mcmc_helpers.R
\docType{data}
\name{X_SAMPLE}
\alias{X_SAMPLE}
\title{Constant used for plotting likelihood functions anf getting mode estimates}
\format{
An object of class \code{numeric} of length 1000.
}
\usage{
X_SAMPLE
}
\description{
Constant used for plotting likelihood functions anf getting mode estimates
}
\keyword{datasets}
